+++ database/migrations/2025_06_05_072931_add_milestone_id_in_project_tempalte_task_table.php
@@ -11,9 +11,9 @@ return new class extends Migration
      */
     public function up(): void
     {
-        if (!Schema::hasColumn('project_template_tasks', 'milestone_id')) {
+        if (Schema::hasTable('project_template_tasks') && !Schema::hasColumn('project_template_tasks', 'milestone_id')) {
             Schema::table('project_template_tasks', function (Blueprint $table) {
-                $table->unsignedInteger('milestone_id')->nullable()->after('project_template_id')->index('tasks_milestone_id_foreign');
+                $table->unsignedInteger('milestone_id')->nullable()->after('project_template_id')->index();
                 $table->foreign(['milestone_id'])->references(['id'])->on('project_template_milestone')->onUpdate('CASCADE')->onDelete('SET NULL');
             });
         }
@@ -24,9 +24,11 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::table('project_template_tasks', function (Blueprint $table) {
-            $table->dropForeign(['milestone_id']);
-            $table->dropColumn('milestone_id');
-        });
+        if (Schema::hasTable('project_template_tasks') && Schema::hasColumn('project_template_tasks', 'milestone_id')) {
+            Schema::table('project_template_tasks', function (Blueprint $table) {
+                $table->dropForeign(['milestone_id']);
+                $table->dropColumn('milestone_id');
+            });
+        }
     }
 };
